+++ Lab8/points.cpp
@@ -1,6 +1,6 @@
 #include "points.h"
 
-Point::Point(int x, int y)
+Point::Point(double x, double y)
 {
     this->x = x;
     this->y = y;
